moved songRender to utils
